Conflicts:
	1503-StructByLightning-LostSamurai-StrawManProject/Lost Samurai/Last Samurai/GameplayState.cpp

+++ 1503-StructByLightning-LostSamurai-StrawManProject/Lost Samurai/Last Samurai/AnimationSystem.h
@@ -27,12 +27,12 @@ public:
 	//Accessor
 	std::map<std::string, Animation>& GetLoaded() { return m_Loaded; }
 	SGD::Rectangle GetRect(const AnimationTimestamp& _info, float _PosX, float _PosY, SGD::Size _scale = { 1, 1 });
-	SGD::Point GetParticlePt(AnimationTimestamp& info, int _X, int _Y, int _index);
+	SGD::Point GetParticlePt(AnimationTimestamp& info, float _X, float _Y, int _index);
 	//SGD::Rectangle GetSonRect() { return m_SonRect; }
 	//Methods
 
 	void Update(int _elapsedtime, AnimationTimestamp& _info);
-	void Render(AnimationTimestamp& _info, int _PosX, int _PosY, SGD::Size _scale = { 1, 1 });
+	void Render(AnimationTimestamp& _info, float _PosX, float _PosY, SGD::Size _scale = { 1, 1 });
 	void Load(const char* _filename);
 	void Exit();
